refactor: create enums
