+++ packages/shared/lib/core/wallet/enums/activity-async-status.enum.ts
@@ -1,0 +1,5 @@
+export enum ActivityAsyncStatus {
+    Unclaimed = 'unclaimed',
+    Claimed = 'claimed',
+    Expired = 'expired',
+}
