+++ packages/shared/lib/core/wallet/enums/activity-direction.enum.ts
@@ -1,0 +1,4 @@
+export enum ActivityDirection {
+    In = 'in',
+    Out = 'out',
+}
