+++ packages/shared/lib/core/wallet/enums/activity-status.enum.ts
@@ -1,0 +1,7 @@
+export enum ActivityStatus {
+    InProgress = 'inProgress',
+    Pending = 'pending',
+    Confirmed = 'confirmed',
+    Conflict = 'conflict',
+    Failed = 'failed',
+}
