+++ packages/shared/lib/core/wallet/enums/activity-type.enum.ts
@@ -1,0 +1,7 @@
+export enum ActivityType {
+    Stake = 'stake',
+    Transfer = 'transfer',
+    Receive = 'receive',
+    Send = 'send',
+    Migrate = 'migrate',
+}
