+++ packages/shared/lib/core/wallet/index.ts
@@ -1,3 +1,4 @@
+export * from './enums'
 export * from './interfaces'
 export * from './stores'
 export * from './utils'
