+++ packages/shared/lib/core/wallet/interfaces/activity.interface.ts
@@ -1,31 +1,5 @@
+import { ActivityType, ActivityDirection } from '../enums'
 import { ITokenMetadata } from './token-metadata.interface'
-
-export enum ActivityType {
-    Stake = 'stake',
-    Transfer = 'transfer',
-    Receive = 'receive',
-    Send = 'send',
-    Migrate = 'migrate',
-}
-
-export enum ActivityStatus {
-    InProgress = 'inProgress',
-    Pending = 'pending',
-    Confirmed = 'confirmed',
-    Conflict = 'conflict',
-    Failed = 'failed',
-}
-
-export enum ActivityAsyncStatus {
-    Unclaimed = 'unclaimed',
-    Claimed = 'claimed',
-    Expired = 'expired',
-}
-
-export enum ActivityDirection {
-    In = 'in',
-    Out = 'out',
-}
 
 export interface IActivity {
     id: string
